+++ common.moplaf.time/com.misc.common.moplaf.time.continuous.emf/src/com/misc/common/moplaf/time/continuous/DistributionVisitor.java
@@ -30,6 +30,6 @@ public interface DistributionVisitor  {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 */
-	boolean visit(Date moment, float amount);
+	boolean visit(Date moment, double amount);
 
 } // DistributionVisitor
