+++ common.moplaf.time/com.misc.common.moplaf.time.continuous.emf/src/com/misc/common/moplaf/time/continuous/impl/AmountAbsoluteImpl.java
@@ -44,7 +44,7 @@ public abstract class AmountAbsoluteImpl extends ProvidedEventImpl implements Am
 	 * @generated
 	 * @ordered
 	 */
-	protected static final float AMOUNT_ABSOLUTE_EDEFAULT = 0.0F;
+	protected static final double AMOUNT_ABSOLUTE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getAmountAbsolute() <em>Amount Absolute</em>}' attribute.
@@ -54,7 +54,7 @@ public abstract class AmountAbsoluteImpl extends ProvidedEventImpl implements Am
 	 * @generated
 	 * @ordered
 	 */
-	protected float amountAbsolute = AMOUNT_ABSOLUTE_EDEFAULT;
+	protected double amountAbsolute = AMOUNT_ABSOLUTE_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -81,7 +81,7 @@ public abstract class AmountAbsoluteImpl extends ProvidedEventImpl implements Am
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public float getAmountAbsolute() {
+	public double getAmountAbsolute() {
 		return amountAbsolute;
 	}
 
@@ -90,12 +90,13 @@ public abstract class AmountAbsoluteImpl extends ProvidedEventImpl implements Am
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setAmountAbsolute(float newAmountAbsolute) {
-		float oldAmountAbsolute = amountAbsolute;
+	public void setAmountAbsolute(double newAmountAbsolute) {
+		double oldAmountAbsolute = amountAbsolute;
 		amountAbsolute = newAmountAbsolute;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, TimeContinuousPackage.AMOUNT_ABSOLUTE__AMOUNT_ABSOLUTE, oldAmountAbsolute, amountAbsolute));
 	}
+
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -132,7 +133,7 @@ public abstract class AmountAbsoluteImpl extends ProvidedEventImpl implements Am
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case TimeContinuousPackage.AMOUNT_ABSOLUTE__AMOUNT_ABSOLUTE:
-				setAmountAbsolute((Float)newValue);
+				setAmountAbsolute((Double)newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
@@ -204,7 +205,7 @@ public abstract class AmountAbsoluteImpl extends ProvidedEventImpl implements Am
 	 */
 	@Override
 	public void refreshAmountAfter() {
-		float amountAfter = this.getAmountAbsolute();
+		double amountAfter = this.getAmountAbsolute();
 		this.setAmountAfter(amountAfter);
 	}
